errore boolean risolto
