+++ funzioni.c
@@ -2,53 +2,28 @@
 
 
 
-void inizializza(stack *stk)
+int main (void)
 {
-  stk->cnt=0;
-  stk->top=NULL;
-}
+  char str[]="mi chiamo joanna kelly!";
+  int i;
+  stack s;
 
-void push (char d,stack *stk)
-{
-  elem *p;
-  p=malloc(sizeof(elem));
-  p->d=d;
-  p->next=stk->top;
-  stk->top=p;
-  stk->cnt++;
-}
+  inizializza(&s);
+  printf("la stringa contiene %s \n",str);
+  for (i=0;str[i]!='\0';i++)
+  {
+    if (!full(&s))
+    {
+      push(str[i],&s);
+    }
+  }
+  printf("dallo stack : \n");
+  while(!empty(&s))
+  {
+    putchar(pop(&s));
+  }
+  putchar('\n');
 
+  return 0;
 
-
-char pop(stack *stk)
-{
-  char r;
-  elem *p;
-
-  r=stk->top->d;
-  p=stk->top;
-  stk->top=stk->top->next;
-  stk->cnt--;
-  free(p);
-  return r;
-}
-
-
-char top(stack *stk)
-{
-  char r;
-  r=stk->top->d;
-  return r;
-}
-boolean empty(stack *stk)
-{
-  int c;
-  c=stk->cnt;
-  return ((boolean) (c==EMPTY));
-}
-boolean full(stack *stk)
-{
-int c;
-c=stk->cnt;
-return ((boolean) (c==FULL));
 }
